Clean up package dependencies (#13)
* Clean up package dependencies

Remove dependencies which were never imported.  Add dependencies which
are imported.  Add @types for modules that have them to
devDependencies.  @types/glob-stream seems incorrect so it was not
added.

Tests pass but compilation does not.

* Fix compile and test errors

Use tmp-promise instead of tmp and fix yargs issue.  Remove dependency
to winston in API.