Add atomist-config and atomist-setup
Add command to create atomist config.  It will prompt for GitHub
credentials and create a GitHub personal access token and write it to
the config.

Add shell script to clone an Atomist API client repo, call
atomist-config, and start up the client, all in one step.  It can be
called from curl.

Add ability to retain simple console log format on CLIs and in tests.