Fix clone token, unify repo creation, move cache
Unify repo creation and move it to gitHub.ts.  Centralize creation of
the clone URL and ensure the token in the clone URL is always viewed
as such by GitHub.

Create directories with fs-extra.ensureDir, it is a lot easier and
more reliable.

Move repo cache to ~/.atomist/cache/repos/github.com/OWNER/REPO.

Closes #63