Separate bitbucket API calls from other tests
because when they are run together, we get a 404 response on repository creation.
I have no idea why and can't duplicate this behavior outside of running all tests.
It happens on travis and Rod's machine and mine.

I know you'll hate this, dd, but I want to get back to work so this gets the weirdness out of the way.