allow parameters in subscriptions to be resolved
This allows to specify the usage of parameters when using `GraqhQL.subscriptionFromFile(path, current, parameters))`

eg. the following graphql file has three parameters:

```
subscription SomeSubscription {
    ChatTeam(id: $teamId) {
        orgs {
            repo(first: 100, offset: $offset, private: $isPrivate) {
                owner
                name
            }
        }
    }
}
```

When using this with

```
const query = GraphQL.subscriptionFromFile(
            "./someSubscription",
            __dirname,
            {
                teamId: "T1L0VDKJP",
                offset: 100,
                isPrivate: true,
            });
```

the variables get replaced in the predicates.

NOTE: this doesn't give you full proper support variables in subscriptions right now but we don't support that in the API anyway as of today. This is a good first step.