Centralize configuration handling into client
Currently it is difficult to use the capabilities available to you in
container runtimes to override the automation configuration.  This
organizes configuration interfaces, renames some with backward
compatibility, and defines a clear hierarchy of configuration sources
that allow for straightforward use of container runtime options like
Kubernetes secrets and Cloud Foundry services.

This also allows the user and module configs to specify pretty much
any automation config value, not just team IDs and token.

Add TypeDoc for configuration interfaces.

These changes are largely backward-compatible.  The exceptions are the
location of the server option interfaces, which most likely are not
referenced outside this package, and restricting teamIds to a string[]
rather than either a string or string[].  The latter will result in a
compile error in clients that use a string and is an easy fix.  The
changes allow automationClient.ts to be simplified.

Also, fix --version command-line option.