Work around Type ... is not assignable to type Configuration'
When Configuration contains a class with private/protected fields, at any level,
then a Configuration that resolves to one installation of automation-client
is _never_ the same type that resolves to a Configuration that resolves o
another installation, even if they're the same version, if they're in different
spots on the filesystem. TypeScript protects us. :-/
So don't do it. It's better to never require a class, but only an interface.
Right now I want to see if this will get us around the problem.