Replace run.ts with start.command.ts
The current run.ts does not work with automation clients that rely on
scanning for command handlers.  Rather than fix that script, which
resides in the now deprecated cli subdirectory, I created
start.command.ts for two reasons.  First, it is easier to determine if
an automation client has a working script to run command handlers, it
is just a check for a file rather than having to parse it.  Second,
the new name parallels the start.client.ts script.