Fail when tests fail, remove team/token
Mocha was not exiting with non-zero status when tests were failing if
the `--exit` command-line option was provided.  Fortunately, we do not
need that anymore and can remove it.  Fix the configuration tests that
were failing.  This includes updating use of teamIds with workspaceIds
and token with apiKey.  Clean up other uses/mentions of these as well.