Make killing processes more cross-platform
Killing processes on win32 was not working.  This was likely due to a
combination of things, including that cross-spawn runs a command shell
when executing certain command and some commands, like mvnw, execute
other commands that ChildProcess.kill() does not reach through to
kill.  On win32, use tree-kill to kill a process and all its children.
Since this does not seem to be necessary on other platforms, we do not
use it.  This logic is in the exported function crossKill.

Add timeout to poisonAndWait and sending SIGKILL after initial
timeout.

Alphabetize index.  Improve TypeDoc.

Hopefully closes #384.