Exclude ingesters from gql-gen
Do not include ingester specifications from GraphQL type generation.
This change also fixes the location of the `graphql` subdirectory to
directly under the `lib` or `src` directory.  This is a change, but
likely not breaking as the location of a project-specific schema.json
was already limited to `{lib,src}/graphql`.