Upgrade tree-path and microgrammar
the goal is to get to microgrammar 1.0.3, with no 1.0.1 on the path, so that
we can allow SDMs to upgrade. 1.0.3 has a fix to remove Concat (class with
a private field) from the Microgrammar type, so that there isn't a conflict
between minor versions of the Microgrammar type.
Until then we get compile errors on any version of Microgrammar except precisely the one client uses.