Update failure logging
Prior to this commit their was always a message that said 'Probably
exists' when a new repo failed to create from a generator.  This was
misleading when the errors where coming from something like an
authorization problem (401).  Updated to make it more generic and let
the `err` object provide the detail.