WebSocket can be destroyed before message sent
There were unhandled promise exceptions in the test because tests were
creating a websocket that never was connected and so was unable to
send messages.  The interval function kept running after the websocket
was destroyed, causing `this` to become undefined.  Since this could
also adversely affect normal operation, I added a check to ensure
`this` was defined.