Rework cluster shutdown WIP
The master organizes shutdown of a cluster, waiting for the queue to
empty and telling the workers to shutdown.  If the workers receive a
shutdown signal, they check and potentially wait to see if the master
has sent them a shutdown message.  If not, they shutdown anyway but at
least give themselves some time to complete work.  Configure max
shutdown timeout based on configured grace period.  Execute all
shutdown hooks, even if some throw an error.

If not shutting down, replace any worker that exits, regardless of
how.

Delint files I touched.

Towards #543