Use major version only for registration
Each registration is being hashed (commands, events, ingesters, name and version) in order to determine if the starting SDM is going to connect to an  existing or a new session.

This means if we keep changing the version number on each deploy or startup, we'll end up with new sessions all the time which is not desirable. We only want a new session when the registration payload (events, commands or ingesters) do actually change.

This change will always only send the major version of the SDM version to keep it stable and allow us to connect to an existing session. If a new session should is desired, the major version can be rev'ed indicating an incompatible change.